Revert "Revert "feat: 랜덤 게임 생성 시 성능 개선""

+++ apps/backend/src/main/java/com/peekle/domain/workbook/dto/response/WorkbookProblemResponse.java
@@ -10,7 +10,7 @@ import lombok.Getter;
 public class WorkbookProblemResponse {
     // 필요한 정보만 골라서, frontend측으로 보내기
     private Long id;
-    private int number;
+    private long number;
     private String title;
     private String tier;
     private String url;
@@ -22,7 +22,7 @@ public class WorkbookProblemResponse {
 
         return WorkbookProblemResponse.builder()
                 .id(problem.getId())
-                .number(Integer.parseInt(problem.getExternalId()))
+                .number(Long.parseLong(problem.getExternalId()))
                 .title(problem.getTitle())
                 .tier(problem.getTier())
                 .url(problem.getUrl())
